refactor(core): Use DI for eventBus code - Part 1 (no-changelog) (#8434)
