+++ packages/cli/src/InternalHooks.ts
@@ -1,5 +1,6 @@
 import { Service } from 'typedi';
 import { snakeCase } from 'change-case';
+import { get as pslGet } from 'psl';
 import type {
 	AuthenticationMethod,
 	ExecutionStatus,
@@ -10,7 +11,15 @@ import type {
 	WorkflowExecuteMode,
 } from 'n8n-workflow';
 import { TelemetryHelpers } from 'n8n-workflow';
-import { get as pslGet } from 'psl';
+import { InstanceSettings } from 'n8n-core';
+
+import { N8N_VERSION } from '@/constants';
+import type { AuthProviderType } from '@db/entities/AuthIdentity';
+import type { GlobalRole, User } from '@db/entities/User';
+import type { ExecutionMetadata } from '@db/entities/ExecutionMetadata';
+import { SharedWorkflowRepository } from '@db/repositories/sharedWorkflow.repository';
+import { MessageEventBus, type EventPayloadWorkflow } from '@/eventbus';
+import { determineFinalExecutionStatus } from '@/executionLifecycleHooks/shared/sharedHookFunctions';
 import type {
 	IDiagnosticInfo,
 	ITelemetryUserDeletionData,
@@ -18,18 +27,9 @@ import type {
 	IExecutionTrackProperties,
 	IWorkflowExecutionDataProcess,
 } from '@/Interfaces';
-import { Telemetry } from '@/telemetry';
-import type { AuthProviderType } from '@db/entities/AuthIdentity';
-import { eventBus } from './eventbus';
 import { EventsService } from '@/services/events.service';
-import type { GlobalRole, User } from '@db/entities/User';
-import { N8N_VERSION } from '@/constants';
 import { NodeTypes } from '@/NodeTypes';
-import type { ExecutionMetadata } from '@db/entities/ExecutionMetadata';
-import { SharedWorkflowRepository } from '@db/repositories/sharedWorkflow.repository';
-import type { EventPayloadWorkflow } from './eventbus/EventMessageClasses/EventMessageWorkflow';
-import { determineFinalExecutionStatus } from './executionLifecycleHooks/shared/sharedHookFunctions';
-import { InstanceSettings } from 'n8n-core';
+import { Telemetry } from '@/telemetry';
 
 function userToPayload(user: User): {
 	userId: string;
@@ -55,6 +55,7 @@ export class InternalHooks {
 		private sharedWorkflowRepository: SharedWorkflowRepository,
 		eventsService: EventsService,
 		private readonly instanceSettings: InstanceSettings,
+		private readonly eventBus: MessageEventBus,
 	) {
 		eventsService.on(
 			'telemetry.onFirstProductionWorkflowSuccess',
@@ -122,7 +123,7 @@ export class InternalHooks {
 	async onWorkflowCreated(user: User, workflow: IWorkflowBase, publicApi: boolean): Promise<void> {
 		const { nodeGraph } = TelemetryHelpers.generateNodesGraph(workflow, this.nodeTypes);
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.workflow.created',
 				payload: {
 					...userToPayload(user),
@@ -141,7 +142,7 @@ export class InternalHooks {
 
 	async onWorkflowDeleted(user: User, workflowId: string, publicApi: boolean): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.workflow.deleted',
 				payload: {
 					...userToPayload(user),
@@ -173,7 +174,7 @@ export class InternalHooks {
 		}
 
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.workflow.updated',
 				payload: {
 					...userToPayload(user),
@@ -201,7 +202,7 @@ export class InternalHooks {
 		nodeName: string,
 	): Promise<void> {
 		const nodeInWorkflow = workflow.nodes.find((node) => node.name === nodeName);
-		void eventBus.sendNodeEvent({
+		void this.eventBus.sendNodeEvent({
 			eventName: 'n8n.node.started',
 			payload: {
 				executionId,
@@ -219,7 +220,7 @@ export class InternalHooks {
 		nodeName: string,
 	): Promise<void> {
 		const nodeInWorkflow = workflow.nodes.find((node) => node.name === nodeName);
-		void eventBus.sendNodeEvent({
+		void this.eventBus.sendNodeEvent({
 			eventName: 'n8n.node.finished',
 			payload: {
 				executionId,
@@ -255,7 +256,7 @@ export class InternalHooks {
 				workflowName: (data as IWorkflowBase).name,
 			};
 		}
-		void eventBus.sendWorkflowEvent({
+		void this.eventBus.sendWorkflowEvent({
 			eventName: 'n8n.workflow.started',
 			payload,
 		});
@@ -277,7 +278,7 @@ export class InternalHooks {
 		} catch {}
 
 		void Promise.all([
-			eventBus.sendWorkflowEvent({
+			this.eventBus.sendWorkflowEvent({
 				eventName: 'n8n.workflow.crashed',
 				payload: {
 					executionId,
@@ -435,11 +436,11 @@ export class InternalHooks {
 		};
 		promises.push(
 			telemetryProperties.success
-				? eventBus.sendWorkflowEvent({
+				? this.eventBus.sendWorkflowEvent({
 						eventName: 'n8n.workflow.success',
 						payload: sharedEventPayload,
 				  })
-				: eventBus.sendWorkflowEvent({
+				: this.eventBus.sendWorkflowEvent({
 						eventName: 'n8n.workflow.failed',
 						payload: {
 							...sharedEventPayload,
@@ -480,7 +481,7 @@ export class InternalHooks {
 		publicApi: boolean;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.deleted',
 				payload: {
 					...userToPayload(userDeletionData.user),
@@ -502,7 +503,7 @@ export class InternalHooks {
 		invitee_role: string;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.invited',
 				payload: {
 					...userToPayload(userInviteData.user),
@@ -537,7 +538,7 @@ export class InternalHooks {
 		public_api: boolean;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.reinvited',
 				payload: {
 					...userToPayload(userReinviteData.user),
@@ -596,7 +597,7 @@ export class InternalHooks {
 
 	async onUserUpdate(userUpdateData: { user: User; fields_changed: string[] }): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.updated',
 				payload: {
 					...userToPayload(userUpdateData.user),
@@ -615,7 +616,7 @@ export class InternalHooks {
 		invitee: User;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.invitation.accepted',
 				payload: {
 					invitee: {
@@ -634,7 +635,7 @@ export class InternalHooks {
 
 	async onUserPasswordResetEmailClick(userPasswordResetData: { user: User }): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.reset',
 				payload: {
 					...userToPayload(userPasswordResetData.user),
@@ -673,7 +674,7 @@ export class InternalHooks {
 
 	async onApiKeyDeleted(apiKeyDeletedData: { user: User; public_api: boolean }): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.api.deleted',
 				payload: {
 					...userToPayload(apiKeyDeletedData.user),
@@ -688,7 +689,7 @@ export class InternalHooks {
 
 	async onApiKeyCreated(apiKeyCreatedData: { user: User; public_api: boolean }): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.api.created',
 				payload: {
 					...userToPayload(apiKeyCreatedData.user),
@@ -703,7 +704,7 @@ export class InternalHooks {
 
 	async onUserPasswordResetRequestClick(userPasswordResetData: { user: User }): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.reset.requested',
 				payload: {
 					...userToPayload(userPasswordResetData.user),
@@ -727,7 +728,7 @@ export class InternalHooks {
 		},
 	): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.signedup',
 				payload: {
 					...userToPayload(user),
@@ -751,7 +752,7 @@ export class InternalHooks {
 		public_api: boolean;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.email.failed',
 				payload: {
 					messageType: failedEmailData.message_type,
@@ -769,7 +770,7 @@ export class InternalHooks {
 		authenticationMethod: AuthenticationMethod;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.login.success',
 				payload: {
 					authenticationMethod: userLoginData.authenticationMethod,
@@ -785,7 +786,7 @@ export class InternalHooks {
 		reason?: string;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.login.failed',
 				payload: {
 					authenticationMethod: userLoginData.authenticationMethod,
@@ -808,7 +809,7 @@ export class InternalHooks {
 		public_api: boolean;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.credentials.created',
 				payload: {
 					...userToPayload(userCreatedCredentialsData.user),
@@ -836,7 +837,7 @@ export class InternalHooks {
 		sharees_removed: number | null;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.user.credentials.shared',
 				payload: {
 					...userToPayload(userSharedCredentialsData.user),
@@ -876,7 +877,7 @@ export class InternalHooks {
 		failure_reason?: string;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.package.installed',
 				payload: {
 					...userToPayload(installationData.user),
@@ -914,7 +915,7 @@ export class InternalHooks {
 		package_author_email?: string;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.package.updated',
 				payload: {
 					...userToPayload(updateData.user),
@@ -947,7 +948,7 @@ export class InternalHooks {
 		package_author_email?: string;
 	}): Promise<void> {
 		void Promise.all([
-			eventBus.sendAuditEvent({
+			this.eventBus.sendAuditEvent({
 				eventName: 'n8n.audit.package.deleted',
 				payload: {
 					...userToPayload(deleteData.user),
