+++ packages/cli/src/eventbus/eventBus.controller.ee.ts
@@ -1,5 +1,15 @@
 import express from 'express';
-import { eventBus } from './MessageEventBus/MessageEventBus';
+import type {
+	MessageEventBusDestinationWebhookOptions,
+	MessageEventBusDestinationOptions,
+} from 'n8n-workflow';
+import { MessageEventBusDestinationTypeNames } from 'n8n-workflow';
+
+import { RestController, Get, Post, Delete, Authorized, RequireGlobalScope } from '@/decorators';
+import { AuthenticatedRequest } from '@/requests';
+import { BadRequestError } from '@/errors/response-errors/bad-request.error';
+
+import { MessageEventBus } from './MessageEventBus/MessageEventBus';
 import {
 	isMessageEventBusDestinationSentryOptions,
 	MessageEventBusDestinationSentry,
@@ -9,16 +19,8 @@ import {
 	MessageEventBusDestinationSyslog,
 } from './MessageEventBusDestination/MessageEventBusDestinationSyslog.ee';
 import { MessageEventBusDestinationWebhook } from './MessageEventBusDestination/MessageEventBusDestinationWebhook.ee';
-import type {
-	MessageEventBusDestinationWebhookOptions,
-	MessageEventBusDestinationOptions,
-} from 'n8n-workflow';
-import { MessageEventBusDestinationTypeNames } from 'n8n-workflow';
-import { RestController, Get, Post, Delete, Authorized, RequireGlobalScope } from '@/decorators';
 import type { MessageEventBusDestination } from './MessageEventBusDestination/MessageEventBusDestination.ee';
-import { AuthenticatedRequest } from '@/requests';
 import { logStreamingLicensedMiddleware } from './middleware/logStreamingEnabled.middleware.ee';
-import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 
 // ----------------------------------------
 // TypeGuards
@@ -53,6 +55,8 @@ const isMessageEventBusDestinationOptions = (
 @Authorized()
 @RestController('/eventbus')
 export class EventBusControllerEE {
+	constructor(private readonly eventBus: MessageEventBus) {}
+
 	// ----------------------------------------
 	// Destinations
 	// ----------------------------------------
@@ -61,9 +65,9 @@ export class EventBusControllerEE {
 	@RequireGlobalScope('eventBusDestination:list')
 	async getDestination(req: express.Request): Promise<MessageEventBusDestinationOptions[]> {
 		if (isWithIdString(req.query)) {
-			return await eventBus.findDestination(req.query.id);
+			return await this.eventBus.findDestination(req.query.id);
 		} else {
-			return await eventBus.findDestination();
+			return await this.eventBus.findDestination();
 		}
 	}
 
@@ -75,22 +79,22 @@ export class EventBusControllerEE {
 			switch (req.body.__type) {
 				case MessageEventBusDestinationTypeNames.sentry:
 					if (isMessageEventBusDestinationSentryOptions(req.body)) {
-						result = await eventBus.addDestination(
-							new MessageEventBusDestinationSentry(eventBus, req.body),
+						result = await this.eventBus.addDestination(
+							new MessageEventBusDestinationSentry(this.eventBus, req.body),
 						);
 					}
 					break;
 				case MessageEventBusDestinationTypeNames.webhook:
 					if (isMessageEventBusDestinationWebhookOptions(req.body)) {
-						result = await eventBus.addDestination(
-							new MessageEventBusDestinationWebhook(eventBus, req.body),
+						result = await this.eventBus.addDestination(
+							new MessageEventBusDestinationWebhook(this.eventBus, req.body),
 						);
 					}
 					break;
 				case MessageEventBusDestinationTypeNames.syslog:
 					if (isMessageEventBusDestinationSyslogOptions(req.body)) {
-						result = await eventBus.addDestination(
-							new MessageEventBusDestinationSyslog(eventBus, req.body),
+						result = await this.eventBus.addDestination(
+							new MessageEventBusDestinationSyslog(this.eventBus, req.body),
 						);
 					}
 					break;
@@ -115,7 +119,7 @@ export class EventBusControllerEE {
 	@RequireGlobalScope('eventBusDestination:test')
 	async sendTestMessage(req: express.Request): Promise<boolean> {
 		if (isWithIdString(req.query)) {
-			return await eventBus.testDestination(req.query.id);
+			return await this.eventBus.testDestination(req.query.id);
 		}
 		return false;
 	}
@@ -124,7 +128,7 @@ export class EventBusControllerEE {
 	@RequireGlobalScope('eventBusDestination:delete')
 	async deleteDestination(req: AuthenticatedRequest) {
 		if (isWithIdString(req.query)) {
-			return await eventBus.removeDestination(req.query.id);
+			return await this.eventBus.removeDestination(req.query.id);
 		} else {
 			throw new BadRequestError('Query is missing id');
 		}
