+++ packages/cli/src/eventbus/eventBus.controller.ts
@@ -1,21 +1,23 @@
 import express from 'express';
+import type { IRunExecutionData } from 'n8n-workflow';
+import { EventMessageTypeNames } from 'n8n-workflow';
+
+import { RestController, Get, Post, Authorized, RequireGlobalScope } from '@/decorators';
+import { BadRequestError } from '@/errors/response-errors/bad-request.error';
+
 import { isEventMessageOptions } from './EventMessageClasses/AbstractEventMessage';
 import { EventMessageGeneric } from './EventMessageClasses/EventMessageGeneric';
 import type { EventMessageWorkflowOptions } from './EventMessageClasses/EventMessageWorkflow';
 import { EventMessageWorkflow } from './EventMessageClasses/EventMessageWorkflow';
 import type { EventMessageReturnMode } from './MessageEventBus/MessageEventBus';
-import { eventBus } from './MessageEventBus/MessageEventBus';
+import { MessageEventBus } from './MessageEventBus/MessageEventBus';
 import type { EventMessageTypes, FailedEventSummary } from './EventMessageClasses';
 import { eventNamesAll } from './EventMessageClasses';
 import type { EventMessageAuditOptions } from './EventMessageClasses/EventMessageAudit';
 import { EventMessageAudit } from './EventMessageClasses/EventMessageAudit';
-import type { IRunExecutionData } from 'n8n-workflow';
-import { EventMessageTypeNames } from 'n8n-workflow';
 import type { EventMessageNodeOptions } from './EventMessageClasses/EventMessageNode';
 import { EventMessageNode } from './EventMessageClasses/EventMessageNode';
-import { recoverExecutionDataFromEventLogMessages } from './MessageEventBus/recoverEvents';
-import { RestController, Get, Post, Authorized, RequireGlobalScope } from '@/decorators';
-import { BadRequestError } from '@/errors/response-errors/bad-request.error';
+import { ExecutionDataRecoveryService } from './executionDataRecovery.service';
 
 // ----------------------------------------
 // TypeGuards
@@ -34,6 +36,11 @@ const isWithQueryString = (candidate: unknown): candidate is { query: string } =
 @Authorized()
 @RestController('/eventbus')
 export class EventBusController {
+	constructor(
+		private readonly eventBus: MessageEventBus,
+		private readonly recoveryService: ExecutionDataRecoveryService,
+	) {}
+
 	// ----------------------------------------
 	// Events
 	// ----------------------------------------
@@ -45,17 +52,17 @@ export class EventBusController {
 		if (isWithQueryString(req.query)) {
 			switch (req.query.query as EventMessageReturnMode) {
 				case 'sent':
-					return await eventBus.getEventsSent();
+					return await this.eventBus.getEventsSent();
 				case 'unsent':
-					return await eventBus.getEventsUnsent();
+					return await this.eventBus.getEventsUnsent();
 				case 'unfinished':
-					return await eventBus.getUnfinishedExecutions();
+					return await this.eventBus.getUnfinishedExecutions();
 				case 'all':
 				default:
-					return await eventBus.getEventsAll();
+					return await this.eventBus.getEventsAll();
 			}
 		} else {
-			return await eventBus.getEventsAll();
+			return await this.eventBus.getEventsAll();
 		}
 	}
 
@@ -63,7 +70,7 @@ export class EventBusController {
 	@RequireGlobalScope('eventBusEvent:list')
 	async getFailedEvents(req: express.Request): Promise<FailedEventSummary[]> {
 		const amount = parseInt(req.query?.amount as string) ?? 5;
-		return await eventBus.getEventsFailed(amount);
+		return await this.eventBus.getEventsFailed(amount);
 	}
 
 	@Get('/execution/:id')
@@ -74,7 +81,7 @@ export class EventBusController {
 			if (req.query?.logHistory) {
 				logHistory = parseInt(req.query.logHistory as string, 10);
 			}
-			return await eventBus.getEventsByExecutionId(req.params.id, logHistory);
+			return await this.eventBus.getEventsByExecutionId(req.params.id, logHistory);
 		}
 		return;
 	}
@@ -86,9 +93,9 @@ export class EventBusController {
 		if (req.params?.id) {
 			const logHistory = parseInt(req.query.logHistory as string, 10) || undefined;
 			const applyToDb = req.query.applyToDb !== undefined ? !!req.query.applyToDb : true;
-			const messages = await eventBus.getEventsByExecutionId(id, logHistory);
+			const messages = await this.eventBus.getEventsByExecutionId(id, logHistory);
 			if (messages.length > 0) {
-				return await recoverExecutionDataFromEventLogMessages(id, messages, applyToDb);
+				return await this.recoveryService.recoverExecutionData(id, messages, applyToDb);
 			}
 		}
 		return;
@@ -113,7 +120,7 @@ export class EventBusController {
 				default:
 					msg = new EventMessageGeneric(req.body);
 			}
-			await eventBus.send(msg);
+			await this.eventBus.send(msg);
 		} else {
 			throw new BadRequestError(
 				'Body is not a serialized EventMessage or eventName does not match format {namespace}.{domain}.{event}',
