+++ packages/cli/src/eventbus/index.ts
@@ -1,1 +1,4 @@
-export { eventBus } from './MessageEventBus/MessageEventBus';
+export { MessageEventBus } from './MessageEventBus/MessageEventBus';
+export { EventMessageTypes } from './EventMessageClasses';
+export { EventPayloadWorkflow } from './EventMessageClasses/EventMessageWorkflow';
+export { METRICS_EVENT_NAME, getLabelsForEvent } from './MessageEventBusDestination/Helpers.ee';
