+++ packages/cli/src/services/metrics.service.ts
@@ -8,12 +8,12 @@ import { Service } from 'typedi';
 import EventEmitter from 'events';
 
 import { CacheService } from '@/services/cache/cache.service';
-import type { EventMessageTypes } from '@/eventbus/EventMessageClasses';
 import {
+	MessageEventBus,
 	METRICS_EVENT_NAME,
 	getLabelsForEvent,
-} from '@/eventbus/MessageEventBusDestination/Helpers.ee';
-import { eventBus } from '@/eventbus';
+	type EventMessageTypes,
+} from '@/eventbus';
 import { Logger } from '@/Logger';
 
 @Service()
@@ -21,6 +21,7 @@ export class MetricsService extends EventEmitter {
 	constructor(
 		private readonly logger: Logger,
 		private readonly cacheService: CacheService,
+		private readonly eventBus: MessageEventBus,
 	) {
 		super();
 	}
@@ -151,7 +152,7 @@ export class MetricsService extends EventEmitter {
 		if (!config.getEnv('endpoints.metrics.includeMessageEventBusMetrics')) {
 			return;
 		}
-		eventBus.on(METRICS_EVENT_NAME, (event: EventMessageTypes) => {
+		this.eventBus.on(METRICS_EVENT_NAME, (event: EventMessageTypes) => {
 			const counter = this.getCounterForEvent(event);
 			if (!counter) return;
 			counter.inc(1);
